add package app.entities and moved the Classes Board and Player into it
