+++ src/main/java/game/app/TicTacToe.java
@@ -2,6 +2,6 @@ package game.app;
 
 public class TicTacToe {
     public static void main(String[] args) {
-        
+
     }
 }
