+++ src/test/java/game/entities/BoardTest.java
@@ -1,5 +1,6 @@
-package game;
+package game.entities;
 
+import game.entities.Board;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
